refactor: split context provider for cleaner code
