+++ src/components/ProtectedRoute.tsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet } from "react-router-dom";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../context/auth/AuthContext";
 
 const ProtectedRoute = ({ roles = [] }: { roles?: string[] }) => {
   const { user, isAuthenticated } = useAuth();
