+++ src/components/ThemeToggle.tsx
@@ -1,6 +1,6 @@
 import { IconButton, Tooltip } from "@mui/material";
 import { LightMode, DarkMode } from "@mui/icons-material";
-import { useTheme } from "../context/ThemeContext";
+import { useTheme } from "../context/theme/ThemeContext";
 
 const ThemeToggle = () => {
   const { mode, toggleTheme } = useTheme();
