--- src/context/AuthContext.tsx
@@ -1,49 +1,0 @@
-import React, { createContext, useContext, useMemo, useState } from "react";
-
-interface User {
-  id: string;
-  name: string;
-  email: string;
-  role: "admin" | "user";
-}
-
-type AuthContextType = {
-  user: User | null;
-  isAuthenticated: boolean;
-  login: (email: string, password: string) => Promise<void>;
-  logout: () => void;
-};
-
-const AuthContext = createContext<AuthContextType | undefined>(undefined);
-
-export const AuthProvider = ({ children }: { children: React.ReactNode }) => {
-  const [user, setUser] = useState<User | null>(null);
-
-  const login = async (email: string, password: string) => {
-    // make api call to the server
-  };
-
-  const logout = () => {
-    setUser(null);
-  };
-
-  const value = useMemo(
-    () => ({
-      user,
-      isAuthenticated: !!user,
-      login,
-      logout,
-    }),
-    [user]
-  );
-
-  return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
-};
-
-export const useAuth = () => {
-  const context = useContext(AuthContext);
-  if (!context) {
-    throw new Error("useAuth must be used within an AuthProvider");
-  }
-  return context;
-};
