+++ src/context/auth/AuthProvider.tsx
@@ -1,0 +1,28 @@
+import { useState, useMemo } from "react";
+import { type User, AuthContext } from "./AuthContext";
+
+export const AuthProvider = ({ children }: { children: React.ReactNode }) => {
+  const [user, setUser] = useState<User | null>(null);
+
+  const login = async (email: string, password: string) => {
+    // make api call to the server
+    // For now, we'll simulate a successful login
+    setUser({ id: "1", name: "John Doe", email, role: "user" });
+  };
+
+  const logout = () => {
+    setUser(null);
+  };
+
+  const value = useMemo(
+    () => ({
+      user,
+      isAuthenticated: !!user,
+      login,
+      logout,
+    }),
+    [user]
+  );
+
+  return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
+};
