+++ src/context/theme/ThemeContext.ts
@@ -1,0 +1,18 @@
+import { createContext, useContext } from "react";
+
+export type ThemeMode = "light" | "dark";
+
+export type ThemeContextType = {
+  mode: ThemeMode;
+  toggleTheme: () => void;
+};
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
+
+export const useTheme = () => {
+  const context = useContext(ThemeContext);
+  if (!context) {
+    throw new Error("useTheme must be used within a ThemeProvider");
+  }
+  return context;
+};
